src/agent-app/requirements.txt
aiohttp
aiosignal
alpaca-py
azure-common
azure-core
azure-identity
azure-mgmt-containerinstance
azure-mgmt-core
azure-storage-blob
ipykernel
ipython
jupyter-black
numpy
openai
pandas
semantic-kernel
finnhub-python
flask
gunicorn
pydantic
flask_pydantic
msal
debugpy
dotenv 